etcd_perf file left after fio test when restoring
This commit deletes "etcd_perf" files that could present if someone has ran a fio test under /var/lib/etcd
like specified in the OpenShift docs about the recommended host practices.

As this file is known as non essential, it could be freely deleted by anticipation by the restore pod when
restoring a cluster to a previous state.

Change-Id: Ib6d9e522bd68fff2aba437fe0ae3ec1bd7260baf
Signed-off-by: Romain Forlot <rforlot@yahoo.com>